+++ R/utils.R
@@ -191,8 +191,9 @@
 
   #checks
   checkfun <- function(x, name="", range=c(-Inf, Inf)) {
-    if (any(x[!is.na(x)] %)(% range))
+    if (any(x[!is.na(x)] < range[1] | x[!is.na(x)] > range[2])) {
       stop(sprintf("Parameter %s is out of range [%.0f, %.0f]", name, range[1], range[2]))
+    }
     invisible(NULL)
   }
   #Coordinates
@@ -220,7 +221,7 @@
   xy <- project(xy, crs)
 
   #LCC4 and LCC2
-  vlcc <- extract(.rlcc, xy)$lcc
+  vlcc <- terra::extract(.rlcc, xy)$lcc
   lcclevs <- c("0"="", "1"="Conif", "2"="Conif", "3"="", "4"="",
                "5"="DecidMixed", "6"="DecidMixed", "7"="", "8"="Open", "9"="",
                "10"="Open", "11"="Open", "12"="Open", "13"="Open", "14"="Wet",
@@ -230,16 +231,16 @@
   levels(lcc2) <- c("Forest", "Forest", "OpenWet", "OpenWet")
 
   #TREE
-  vtree <- extract(.rtree, xy)$tree
+  vtree <- terra::extract(.rtree, xy)$tree
   TREE <- vtree / 100
-  TREE[TREE %)(% c(0, 1)] <- 0
+  TREE[TREE < 0 | TREE > 1] <- 0
 
   #raster::extract seedgrow value (this is rounded)
-  d1 <- extract(.rd1, xy)$seedgrow
+  d1 <- terra::extract(.rd1, xy)$seedgrow
 
   #UTC offset + 7 makes Alberta 0 (MDT offset) for local times
   if(tz=="local"){
-    ltz <- extract(.rtz, xy)$utcoffset + 7
+    ltz <- terra::extract(.rtz, xy)$utcoffset + 7
   }
   if(tz=="utc"){
     ltz <- 0